fix: Add kafka mock for unit tests
