+++ apps/srv/certification-service/src/app/certificate/certificate.schema.ts
@@ -5,14 +5,14 @@ import * as toJson from '@meanie/mongoose-to-json';
 
 @Schema()
 export class Certificate extends Document {
-    @Prop({ required: true, unique: true })
-    name: string;
+  @Prop({ required: true, unique: true })
+  name: string;
 
-    @Prop({ required: true })
-    version: string;
+  @Prop({ required: true })
+  version: string;
 
-    @Prop({ required: true })
-    signature: string;
+  @Prop({ required: true })
+  signature: string;
 }
 
 export const CertificateSchema = SchemaFactory.createForClass(Certificate);
