+++ libs/srv/kafka/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/kafka.module';
 export * from './lib/kafka.decorator';
 export * from './lib/event.dto';
 export * from './lib/kafka-exception.filter';
+export * from './lib/kafka.module.mock';
